third_party/WebKit/LayoutTests/http/tests/websocket/accept-language-header.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/get-request-header.js"></script>
<script>description('An Accept-Language header must be present in the request.');

window.jsTestIsAsync = true;

var accept_language_value;

connectAndGetRequestHeader('accept-language').then(function(value)
{
    accept_language_value = value;
    shouldBeNonNull('accept_language_value');
    shouldBeDefined('accept_language_value');
    shouldNotBe('accept_language_value', '""');
    finishJSTest();
}, finishAsFailed);
</script>
