third_party/blink/web_tests/platform/generic/http/tests/websocket/accept-language-header-expected.txt

An Accept-Language header must be present in the request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS accept_language_value is non-null.
PASS accept_language_value is defined.
PASS accept_language_value is not ""
PASS successfullyParsed is true

TEST COMPLETE
